chore(files): rename 'words-ending-utils.ts' -> 'word.ts'
