+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from './main';
 export * from './case';
-export * from './word-ending-utils';
+export * from './word';
