+++ src/word.ts
